new: inital serial, vga driver, font renderer, and mboot2 Now, if VBE is available, then Paging::init maps the framebuffer address to the virtual memory, and then the vga module is inited. Terminal now supports outputting to both Serial and VGA, and by default, chooses Serial output to begin with. If VBE flag is not set, the terminal continues outputting to the serial.
The VGA driver is pretty basic atm. It supports drawing pixels and a
horizontal straight line. The rendering code is also not much efficient.

There is also initial support for fonts, albeit not in any kind of
standardized way. tools/generate_font_def.py extracts ascii glyphs
of a particular size from a particular font, and saves them as a Font
struct. The font renderer uses these preextracted glyphs, which contains
greyscale white value for each pixel, and just plants them onto the
framebuffer. It also supports color interpolation based on the intensity
of the pixel in the original rendering output.

Since not much is changed except for the backend, the frontend terminal
code works as usual.

The terminal does not yet support screen scrolling on VGA. It also does
not support variable width glyphs, and assumes every character to be of
the max width of all characters. These will be addressed next.

Also, we now need to compile with c++17 due to some weird redundant
declaration rule for static constexpr members of a class in std < c++17.